- (auto_nlist.c, auto_nlist.h, kernel.c, snmp_agent.c, snmpd.c,   hr_device.c, hr_disk.c, hr_filesys.c, hr_network.c, hr_other.c,   hr_partition.c, hr_print.c, hr_proc.c, hr_storage.c, hr_swinst.c,   hr_swrun.c, hr_system.c, hr_utils.c, at.c, interfaces.c,   var_route.c, snmp_parse_args.c, snmpdelta.c, snmptable.c,   snmptranslate.c, snmptrapd.c, freebsd2.h, hpux.h, linux.h, netbsd.h,   solaris.h, sysv.h, asn1.c, mib.c, parse.c, parse.h, snmp_auth.c,   snmp_impl.h): Patch from Niels Baggensen:   - make host resources work under NetBSD, FreeBSD, Solaris and Linux.     - SunOS4 almost works.   - Add -s/-S to command line options for printing of shorter oids.   - numerous bug fixes.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@1433 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/snmp_auth.c
@@ -379,7 +379,7 @@ snmp_comstr_build(data, length, sid, slen, version, messagelen)
     int         *length;
     u_char      *sid;
     int         *slen;
-    int         *version;
+    long        *version;
     int         messagelen;
 {
   /* version is an 'int' (CMU had it as a long, but was passing in a *int.
