+++ snmplib/snmp_impl.h
@@ -164,7 +164,7 @@ extern int snmp_errno;
 #define FIRST_PASS	1
 #define	LAST_PASS	2
 u_char	*snmp_comstr_parse __P((u_char *, int *, u_char *, int *, long *));
-u_char	*snmp_comstr_build __P((u_char *, int *, u_char *, int *, int *, int));
+u_char	*snmp_comstr_build __P((u_char *, int *, u_char *, int *, long *, int));
 
 u_char	*snmp_party_parse __P((u_char *, int *, struct packet_info *,
                                  oid *, int *, oid *, int *, oid *, int *, int));
